+++ docs/src/dev/report_gen.md
@@ -14,34 +14,13 @@ The test cases we use for this analysis are:
     - `ark_analytic_nonlin`, which uses a nonlinear `T_imp!`
     - `1d_heat_equation` and `2d_heat_equation`, which use a nonlinear `T_exp!` and `dss!`, where the spatial discretization is implemented using `ClimaCore`
 
-```@example
-include("report_gen.jl")
-```
- ![](output/convergence_ark_analytic_all_algorithms.png)
- ![](output/convergence_ark_analytic_sys_all_algorithms.png)
- ![](output/convergence_ark_onewaycouple_mri_all_algorithms.png)
- ![](output/convergence_ark_analytic_nonlin_all_algorithms.png)
- ![](output/convergence_1d_heat_equation_all_algorithms.png)
- ![](output/convergence_2d_heat_equation_all_algorithms.png)
+Please see the `Summaries` section of our [buildkite results](https://buildkite.com/clima/climatimesteppers-ci/), which has a comprehensive report.
 
 ## Errors with a Limiter
 
 In order to verify the correctness of our algorithms with a limiter, we recreate Table 1 from ["Optimization-based limiters for the spectral element method" by Guba et al.](https://www.sciencedirect.com/science/article/pii/S0021999114001491) This involves running the `horizontal_deformational_flow` test case (from ["A standard test case suite for two-dimensional linear transport on the sphere" by Lauritzen et al.](https://gmd.copernicus.org/articles/5/887/2012/gmd-5-887-2012.pdf)) with and without a limiter, and also with and without hyperdiffusion. This test case uses a limited tendency `T_lim!` (which consists of advection and, optionally, hyperdiffusion), along with `dss!` and `lim!`. The spatial discretization is implemented using `ClimaCore`. Since this analysis is relatively expensive to run, we only check the results for `SSP333` and `ARS343`. Note that it is possible to limit undershoots and overshoots to 0 (up to floating-point roundoff error) when using the `SSP` `SSP333`, but not when using the `Unconstrained` `ARS343`.
 
-```@example
-using ClimaTimeSteppers # hide
-ENV["GKSwstype"] = "nul" # hide
-include(joinpath(@__DIR__, "..", "plotting_utils.jl")) # hide
-include(joinpath(pkgdir(ClimaTimeSteppers), "test", "problems.jl")) # hide
-
-# This also runs with num_steps = 1000, but with larger under/overshoots; 4800
-# is the value used in the paper.
-limiter_summary(Float64, [SSP333(), ARS343()], horizontal_deformational_flow_test, 4800)
-```
-
-Plots of the tracer specific humidities that were used to compute this table are shown below.
- ![](output/limiter_summary_SSP333.png)
- ![](output/limiter_summary_ARS343.png)
+Please see the `Summaries` section of our [buildkite results](https://buildkite.com/clima/climatimesteppers-ci/), which has a comprehensive report.
 
 ## References
