+++ docs/src/plotting_utils.jl
@@ -99,265 +99,15 @@ function make_saving_callback(cb, u, t, integrator)
     return DiffEqCallbacks.SavingCallback(cb, DiffEqCallbacks.SavedValues(typeof(t), savevalType))
 end
 
-function verify_convergence(
-    title,
-    algorithm_names,
-    test_case,
-    num_steps;
-    num_steps_scaling_factor = 10,
-    order_confidence_percent = 99,
-    super_convergence = (),
-    numerical_reference_algorithm_name = nothing,
-    numerical_reference_num_steps = num_steps_scaling_factor^3 * num_steps,
-    full_history_algorithm_name = nothing,
-    average_function = array -> norm(array) / sqrt(length(array)),
-    average_function_str = "RMS",
-    only_endpoints = false,
-    verbose = false,
-)
-    (; test_name, t_end, linear_implicit, analytic_sol) = test_case
-    prob = test_case.split_prob
-    FT = typeof(t_end)
-    default_dt = t_end / num_steps
-
-    algorithm(algorithm_name::ClimaTimeSteppers.ERKAlgorithmName) = ExplicitAlgorithm(algorithm_name)
-    algorithm(algorithm_name::ClimaTimeSteppers.SSPKnoth) =
-        ClimaTimeSteppers.RosenbrockAlgorithm(ClimaTimeSteppers.tableau(ClimaTimeSteppers.SSPKnoth()))
-    algorithm(algorithm_name::ClimaTimeSteppers.IMEXARKAlgorithmName) =
-        IMEXAlgorithm(algorithm_name, NewtonsMethod(; max_iters = linear_implicit ? 1 : 2))
-
-    ref_sol = if isnothing(numerical_reference_algorithm_name)
-        analytic_sol
-    else
-        ref_alg = algorithm(numerical_reference_algorithm_name)
-        ref_alg_str = string(nameof(typeof(numerical_reference_algorithm_name)))
-        ref_dt = t_end / numerical_reference_num_steps
-        verbose &&
-            @info "Generating numerical reference solution for $test_name with $ref_alg_str (dt = $ref_dt)..."
-        solve(deepcopy(prob), ref_alg; dt = ref_dt, save_everystep = !only_endpoints)
-    end
-
-    cur_avg_err(u, t, integrator) = average_function(abs.(u .- ref_sol(t)))
-    cur_avg_sol_and_err(u, t, integrator) = (average_function(u), average_function(abs.(u .- ref_sol(t))))
-
-    float_str(x) = @sprintf "%.4f" x
-    pow_str(x) = "10^{$(@sprintf "%.1f" log10(x))}"
-    function si_str(x)
-        x in (0, Inf, -Inf, NaN) && return string(x)
-        exponent = floor(Int, log10(x))
-        mantissa = x / 10.0^exponent
-        return "$(float_str(mantissa)) \\times 10^{$exponent}"
-    end
-
-    net_avg_sol_str = "\\textrm{$average_function_str}\\_\\textrm{solution}"
-    net_avg_err_str = "\\textrm{$average_function_str}\\_\\textrm{error}"
-    cur_avg_sol_str = "\\textrm{current}\\_$net_avg_sol_str"
-    cur_avg_err_str = "\\textrm{current}\\_$net_avg_err_str"
-
-    linestyles = (:solid, :dash, :dot, :dashdot, :dashdotdot)
-    marker_kwargs = (; markershape = :circle, markeralpha = 0.5, markerstrokewidth = 0)
-    plot_kwargs = (;
-        legendposition = :outerright,
-        legendtitlefontpointsize = 8,
-        palette = :glasbey_bw_minc_20_maxl_70_n256,
-        size = (1000, 2000), # size in px
-        leftmargin = 60Plots.px,
-        rightmargin = 0Plots.px,
-        topmargin = 0Plots.px,
-        bottommargin = 30Plots.px,
-    )
-
-    plot1_dts = t_end ./ round.(Int, num_steps .* num_steps_scaling_factor .^ (-1:0.5:1))
-    plot1 = Plots.plot(;
-        title = "Convergence Orders",
-        xaxis = (latexstring("dt"), :log10),
-        yaxis = (latexstring(net_avg_err_str), :log10),
-        legendtitle = "Convergence Order ($order_confidence_percent% CI)",
-        plot_kwargs...,
-    )
-
-    plot2b_min = typemax(FT)
-    plot2b_max = typemin(FT)
-    plot2a = Plots.plot(;
-        title = latexstring("Solutions with \$dt = $(pow_str(default_dt))\$"),
-        xaxis = (latexstring("t"),),
-        yaxis = (latexstring(cur_avg_sol_str),),
-        legendtitle = latexstring(net_avg_sol_str),
-        plot_kwargs...,
-    )
-    plot2b = Plots.plot(;
-        title = latexstring("Errors with \$dt = $(pow_str(default_dt))\$"),
-        xaxis = (latexstring("t"),),
-        yaxis = (latexstring(cur_avg_err_str), :log10),
-        legendtitle = latexstring(net_avg_err_str),
-        plot_kwargs...,
-    )
-
-    scb_cur_avg_err = make_saving_callback(cur_avg_err, prob.u0, t_end, nothing)
-    scb_cur_avg_sol_and_err = make_saving_callback(cur_avg_sol_and_err, prob.u0, t_end, nothing)
-
-    for algorithm_name in algorithm_names
-        alg = algorithm(algorithm_name)
-        alg_str = string(nameof(typeof(algorithm_name)))
-        predicted_order = predicted_convergence_order(algorithm_name, prob.f)
-        linestyle = linestyles[(predicted_order - 1) % length(linestyles) + 1]
-
-        verbose && @info "Running $test_name with $alg_str..."
-        plot1_net_avg_errs = map(plot1_dts) do plot1_dt
-            cur_avg_errs =
-                solve(
-                    deepcopy(prob),
-                    alg;
-                    dt = plot1_dt,
-                    save_everystep = !only_endpoints,
-                    callback = scb_cur_avg_err,
-                ).u
-            verbose && @info "RMS_error(dt = $plot1_dt) = $(average_function(cur_avg_errs))"
-            return average_function(cur_avg_errs)
-        end
-        order, order_uncertainty = convergence_order(plot1_dts, plot1_net_avg_errs, order_confidence_percent / 100)
-        order_str = "$(float_str(order)) \\pm $(float_str(order_uncertainty))"
-        if algorithm_name in super_convergence
-            predicted_order += 1
-            plot1_label = "$alg_str: \$$order_str\\ \\ \\ \\textbf{\\textit{SC}}\$"
-        else
-            plot1_label = "$alg_str: \$$order_str\$"
-        end
-        verbose && @info "Order = $order ± $order_uncertainty"
-        if abs(order - predicted_order) > order_uncertainty
-            @warn "Predicted order outside error bars for $alg_str ($test_name)"
-        end
-        if order_uncertainty > predicted_order / 10
-            @warn "Order uncertainty too large for $alg_str ($test_name)"
-        end
-        Plots.plot!(plot1, plot1_dts, plot1_net_avg_errs; label = latexstring(plot1_label), linestyle, marker_kwargs...)
-
-        # Remove all 0s from plot2_cur_avg_errs because they cannot be plotted on a
-        # logarithmic scale. Record the extrema of plot2_cur_avg_errs to set ylim.
-        plot2_values = solve(
-            deepcopy(prob),
-            alg;
-            dt = default_dt,
-            save_everystep = !only_endpoints,
-            callback = scb_cur_avg_sol_and_err,
-        )
-        plot2_ts = plot2_values.t
-        plot2_cur_avg_sols = first.(plot2_values.u)
-        plot2_cur_avg_errs = last.(plot2_values.u)
-        plot2b_min = min(plot2b_min, minimum(x -> x == 0 ? typemax(FT) : x, plot2_cur_avg_errs))
-        plot2b_max = max(plot2b_max, maximum(plot2_cur_avg_errs))
-        plot2_cur_avg_errs .= max.(plot2_cur_avg_errs, eps(FT(0)))
-        plot2_net_avg_sol = average_function(plot2_cur_avg_sols)
-        plot2_net_avg_err = average_function(plot2_cur_avg_errs)
-        Plots.plot!(
-            plot2a,
-            plot2_ts,
-            plot2_cur_avg_sols;
-            label = latexstring("$alg_str: \$$(si_str(plot2_net_avg_sol))\$"),
-            linestyle,
-            (only_endpoints ? marker_kwargs : (;))...,
-        )
-        Plots.plot!(
-            plot2b,
-            plot2_ts,
-            plot2_cur_avg_errs;
-            label = latexstring("$alg_str: \$$(si_str(plot2_net_avg_err))\$"),
-            linestyle,
-            (only_endpoints ? marker_kwargs : (;))...,
-        )
-    end
-
-    Plots.plot!(plot2b; ylim = (plot2b_min / 2, plot2b_max * 2))
-
-    plots = (plot1, plot2a, plot2b)
-
-    if !isnothing(full_history_algorithm_name)
-        history_alg = algorithm(full_history_algorithm_name)
-        history_alg_name = string(nameof(typeof(full_history_algorithm_name)))
-        history_solve_results = solve(
-            deepcopy(prob),
-            history_alg;
-            dt = default_dt,
-            save_everystep = !only_endpoints,
-            callback = make_saving_callback((u, t, integrator) -> u .- ref_sol(t), prob.u0, t_end, nothing),
-        )
-        history_array = hcat(history_solve_results.u...)
-        history_plot_title = "Errors for $history_alg_name with \$dt = $(pow_str(default_dt))\$"
-        history_plot = Plots.heatmap(
-            history_solve_results.t,
-            1:size(history_array, 1),
-            history_array;
-            title = latexstring(history_plot_title),
-            xaxis = (latexstring("t"),),
-            yaxis = ("index",),
-            plot_kwargs...,
-        )
-        plots = (plots..., history_plot)
-    end
-
-    avg_def_str(val_str, var_str) = "\\textrm{$average_function_str}(\\{$val_str\\,\\forall\\,$var_str\\})"
-    ref_sol_def_str =
-        isnothing(numerical_reference_algorithm_name) ? "Y_{analytic}(t)[\\textrm{index}]" :
-        "Y_{ref}(t)[\\textrm{index}]"
-    cur_avg_sol_def_str = avg_def_str("Y(t)[\\textrm{index}]", "\\textrm{index}")
-    cur_avg_err_def_str = avg_def_str("|Y(t)[\\textrm{index}] - $ref_sol_def_str|", "\\textrm{index}")
-    net_avg_sol_def_str = avg_def_str("$cur_avg_sol_str(t)", "t")
-    net_avg_err_def_str = avg_def_str("$cur_avg_err_str(t)", "t")
-    footnote = """
-        Terminology:
-            \$$cur_avg_sol_str(t) = $cur_avg_sol_def_str\$
-            \$$cur_avg_err_str(t) = $cur_avg_err_def_str\$
-            \$$net_avg_sol_str = $net_avg_sol_def_str\$
-            \$$net_avg_err_str = $net_avg_err_def_str\$"""
-    if !isnothing(numerical_reference_algorithm_name)
-        ref_cur_avg_errs = map(ref_sol.u, ref_sol.t) do u, t
-            average_function(abs.(u .- analytic_sol(t)))
-        end
-        ref_net_avg_err = average_function(ref_cur_avg_errs)
-        ref_cur_avg_err_def_str =
-            avg_def_str("|Y_{ref}(t)[\\textrm{index}] - Y_{analytic}(t)[\\textrm{index}]|", "\\textrm{index}")
-        ref_net_avg_err_def_str = avg_def_str(ref_cur_avg_err_def_str, "t")
-        footnote = "$footnote\n\n\nNote: The \"reference solution\" \$Y_{ref}\$ was \
-                    computed using $ref_alg_str with\n\$dt = $(pow_str(ref_dt)),\\ \
-                    \\textrm{and}\\ $ref_net_avg_err_def_str = \
-                    $(si_str(ref_net_avg_err))\$"
-    end
-    footnote_plot = Plots.plot(;
-        title = latexstring(footnote),
-        titlelocation = :left,
-        titlefontsize = 12,
-        axis = nothing,
-        framestyle = :none,
-        margin = 0Plots.px,
-        bottommargin = 10Plots.px,
-    )
-    plots = (plots..., footnote_plot)
-
-    n_plots = length(plots)
-    heights₀ = repeat([1 / (n_plots - 1)], n_plots - 1)
-    footnote_height = eps()
-    heights = [heights₀ .- footnote_height / n_plots..., footnote_height]
-    layout = Plots.grid(n_plots, 1; heights)
-    plot = Plots.plot(
-        plots...;
-        plot_title = "Analysis of $title for \"$test_name\"",
-        layout = layout,
-        fontfamily = "Computer Modern",
-    )
-
-    mkpath("output")
-    file_suffix = lowercase(replace(test_name * ' ' * title, " " => "_"))
-    Plots.savefig(plot, joinpath("output", "convergence_$file_suffix.png"))
-end
-
 # Generates the Table 1 from
 # "Optimization-based limiters for the spectral element method" by Guba et al.,
 # and also plots the values used to generate the table.
-function limiter_summary(::Type{FT}, algorithm_names, test_case_type, num_steps) where {FT}
+function limiter_summary(sol_dicts, algorithm_names)
     to_title(name) = titlecase(replace(string(name), '_' => ' '))
     table_rows = []
-    for algorithm_name in algorithm_names
-        alg_str = string(nameof(typeof(algorithm_name)))
+    mkpath("output")
+    for alg_name in algorithm_names
+        alg_str = string(nameof(typeof(alg_name)))
         plots = []
         plot_kwargs = (;
             clims = (0, 1),
@@ -367,13 +117,7 @@ function limiter_summary(::Type{FT}, algorithm_names, test_case_type, num_steps)
             margin = 10Plots.px,
         )
         for use_limiter in (false, true), use_hyperdiffusion in (false, true)
-            test_case = test_case_type(FT; use_limiter, use_hyperdiffusion)
-            prob = test_case.split_prob
-            dt = test_case.t_end / num_steps
-            algorithm =
-                algorithm_name isa ClimaTimeSteppers.IMEXARKAlgorithmName ?
-                IMEXAlgorithm(algorithm_name, NewtonsMethod()) : ExplicitAlgorithm(algorithm_name)
-            solution = solve(deepcopy(prob), algorithm; dt).u
+            solution = sol_dicts[alg_name, use_hyperdiffusion, use_limiter]
             initial_q = solution[1].ρq ./ solution[1].ρ
             final_q = solution[end].ρq ./ solution[end].ρ
             names = propertynames(initial_q)
